BE-22: change ID type

+++ src/main/java/az/gdg/mssubscriber/repository/SubscriberRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubscriberRepository extends JpaRepository<SubscriberEntity, Integer> {
+public interface SubscriberRepository extends JpaRepository<SubscriberEntity, Long> {
     SubscriberEntity findSubscriberByEmail(String email);
 
 }
